feat: Added basic user model and create endpoint
This model is mostly complete for now. It is just missing some way to do
authentication as well as authorization. Those are coming next.

As with the networks, this is all being done in memory for now and the
actual persistence layer is coming.

This does create an issue that will need to be addressed when
authentication is added. How is the first user going to be created when
authentication is required? There is a possibility of having a setup
endpoint that can be used as a one-off.